RocketMQ extension Merged into one
