+++ trace-etl/trace-etl-api/src/main/java/com/xiaomi/hera/trace/etl/bo/MqConfig.java
@@ -3,6 +3,7 @@ package com.xiaomi.hera.trace.etl.bo;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.util.List;
 import java.util.function.Function;
 
 /**
@@ -13,12 +14,15 @@ import java.util.function.Function;
 public class MqConfig<T> implements Serializable {
 
 
-    private String group;
-
     private String nameSerAddr;
 
-    private String topicName;
+    private String consumerGroup;
+    private String consumerTopicName;
+
+    private String producerGroup;
+    private String producerTopicName;
 
     private Function<T, Boolean> consumerMethod;
+    private Function<List<T>, Boolean> batchConsumerMethod;
 
 }
