+++ trace-etl/trace-etl-server/src/main/resources/application.properties
@@ -43,4 +43,4 @@ mq.rocketmq.es.topic=${mq.rocketmq.es.topic}
 mq.rocketmq.producer.group=${mq.rocketmq.producer.group}
 mq.rocketmq.consumer.group=${mq.rocketmq.consumer.group}
 
-extensions=mqConsumer:rocketMQConsumer:run.mone.trace.etl.extension.rocketmq.consumer,mqProducer:rocketMQProducer:run.mone.trace.etl.extension.rocketmq.producer
+extensions=mq:rocketMQ:run.mone.trace.etl.extension.rocketmq
